% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Candidate_genes_specificity.R
\name{Candidate_genes_specificity}
\alias{Candidate_genes_specificity}
\title{Candidate genes specificity}
\usage{
Candidate_genes_specificity(exp, annotation, candidate, correlation)
}
\arguments{
\item{exp}{A Seurat format single-cell expression matrix}

\item{annotation}{Cell type annotation column names in seurat meta.data part}

\item{candidate}{Candidate gene list}

\item{correlation}{Correlation lists of genes and phenotypes with colnames 'gene' and 'cor'}
}
\value{
A list with specificity percent and the rank of specificity score and correlation for candidate genes
}
\description{
Calculate and sequence the percentage of candidate gene expression in each cell type
}
\details{
Input a Seurat format single-cell expression matrix, cell type annotation column names, candidate gene list, and correlation lists of genes and phenotypes
}
\author{
Yijun Zhou
}
