% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDrugHunter-main.R
\name{scDGS}
\alias{scDGS}
\title{Single Cell Drug Hunter}
\usage{
scDGS(seurat_obj, gene_cor, gene_pvalue, candidate)
}
\arguments{
\item{seurat_obj}{A Seurat format single-cell expression matrix with 'Cell_type' colname in meta.data}

\item{gene_cor}{Correlation lists of genes and phenotypes with colnames 'gene' and 'cor'}

\item{gene_pvalue}{P-value lists of genes and phenotypes with colnames 'gene' and 'pvalue'}

\item{candidate}{Candidate gene list}
}
\value{
A list with scDGS and the rank of specificity score, correlation, drug-gene interaction and pvalue for candidate drug-gene pairs
}
\description{
Calculate single cell Drug Gene Score(scDGS) for candidate genes' drug-gene pairs based on DGIdb 2021.5 database
}
\details{
Input a Seurat format single-cell expression matrix, cell type annotation column names, candidate gene list, correlation lists of genes and phenotypes, significance pvalue for genes
}
\author{
Yijun Zhou
}
