% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDH_circlize_plot.R
\name{scDGS_circlize_plot}
\alias{scDGS_circlize_plot}
\title{scDH_circlize_plot}
\usage{
scDGS_circlize_plot(DG_pairs, annotation, cell_type, top_g, top_d)
}
\arguments{
\item{DG_pairs}{A list of scDH results}

\item{annotation}{Cell type annotation column names in scDH results}

\item{cell_type}{Candidate cell type}

\item{top_g}{Number of top genes to be ploted in circlize}

\item{top_d}{Number of top drugs for gene specific drug_gene pairs to be ploted in circlize}
}
\value{
A circlize chart for top drug-gene pairs in candidate cell type
}
\description{
Plot the top drug-gene pairs' scDGS network in specific cell type
}
\details{
Input scDH format results
}
\author{
Yijun Zhou
}
