% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDH_radar_plot.R
\name{scDH_radar_plot}
\alias{scDH_radar_plot}
\title{scDH_radar_plot}
\usage{
scDH_radar_plot(DG_pairs, annotation, cell_type, top)
}
\arguments{
\item{DG_pairs}{A list of scDH results}

\item{annotation}{Cell type annotation column names in scDH results}

\item{cell_type}{Candidate cell type}

\item{top}{Number of top pairs to be ploted in radar chart}
}
\value{
A radar chart for top n drug-gene pairs in candidate cell type
}
\description{
Plot the top n drug-gene pairs' rank in specific cell type
}
\details{
Input scDGS format results
}
\author{
Yijun Zhou
}
